use (int) for framerate
git-svn-id: https://pd-gem.svn.sourceforge.net/svnroot/pd-gem/trunk/Gem@2375 49b2a0ca-c541-4d49-a7ef-d785e4371512

+++ src/Pixes/recordQT4L.h
@@ -86,7 +86,7 @@ class GEM_EXTERN recordQT4L : public record {
   // framedur is the duration of one frame in [ms]
   //   
   // returns TRUE if init was successfull, FALSE otherwise 
-  virtual bool init(const imageStruct* dummyImage, const float framedur);
+  virtual bool init(const imageStruct* dummyImage, const int framedur);
 
 
   //////////
